+++ src/distributions.hpp
@@ -9,7 +9,7 @@
 
 double normalCDF(double x, double mu, double stddev);
 double sampleY();
-int poisson(double theta);
+double poisson(double theta);
 double gamma(double a);
 double chi_square(double v);
